% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_classes.R
\name{add_lipid_class}
\alias{add_lipid_class}
\title{Add custom lipid class to \code{lipid_classes}}
\usage{
add_lipid_class(lipid_class)
}
\arguments{
\item{lipid_class}{string. Lipid class abbreviation}
}
\value{
\code{lipid_classes} vector + lipid_class-argument
}
\description{
The \code{add_lipid_class()} function is used to extend the
\code{lipid_classes} vector. This is necessary, if custom lipid classes
needs to be evaluated and are not in the \code{lipid_classes} vector.
}
\examples{
lipid_classes <- add_lipid_class("SGalCer")
}
