% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lipid_classes.R
\docType{data}
\name{lipid_classes}
\alias{lipid_classes}
\title{Vector with lipid class abbreviations}
\format{
A \code{vector} with 59 elements
}
\usage{
lipid_classes
}
\description{
A \code{vector} containing lipid classes that is for parsing
titles of LipidView text files. This vector will be used by function
\code{name_lipid_file()}.
}
\keyword{datasets}
