% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipidview_files.R
\name{list_txt_files}
\alias{list_txt_files}
\title{List files with '.txt'-pattern from given path and set names of each object}
\usage{
list_txt_files(path)
}
\arguments{
\item{path}{string. Path to your textfiles}
}
\value{
character vector
}
\description{
List files with '.txt'-pattern from given path and set names of each object
}
\examples{
list_txt_files(system.file("extdata", package = "lipidcountr"))
}
