% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipidview_files.R
\name{name_lipid_file}
\alias{name_lipid_file}
\title{Helper function: Name a single txt.file}
\usage{
name_lipid_file(file, names = lipid_classes)
}
\arguments{
\item{file}{character. Name/path of textfile}

\item{names}{character vector. Vector of lipidclasses to parse out of your files}
}
\value{
character. Named version of name/path of your textfile
}
\description{
Helper function: Name a single txt.file
}
