% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_lipidview_files.R
\name{read_lipidview_files}
\alias{read_lipidview_files}
\title{Read, merge and clean txt.files, which are generated by LipidView}
\usage{
read_lipidview_files(path, format = "long")
}
\arguments{
\item{path}{string. Path to your textfiles}

\item{format}{string
\itemize{
\item "long": Default. Prints the tibble in long-format (all samples in one col)
\item "wide": Prints the tibble in wide-format (one sample, one col)
}}
}
\value{
tibble. Merged textfiles into one tibble
}
\description{
Read, merge and clean txt.files, which are generated by LipidView
}
\examples{
read_lipidview_files(system.file("extdata", package = "lipidcountr"))
}
